임식 커밋, API 2개 구현 완료

+++ src/main/java/UMC_7th_Hackathon_M_Team/demo/domain/gameMember/entity/GameMember.java
@@ -1,6 +1,6 @@
 package UMC_7th_Hackathon_M_Team.demo.domain.gameMember.entity;
 
-import UMC_7th_Hackathon_M_Team.demo.domain.group.entity.Team;
+import UMC_7th_Hackathon_M_Team.demo.domain.Team.entity.Team;
 import UMC_7th_Hackathon_M_Team.demo.domain.member.entity.Member;
 import UMC_7th_Hackathon_M_Team.demo.global.common.BaseEntity;
 import jakarta.persistence.Column;
@@ -31,7 +31,7 @@ public class GameMember extends BaseEntity {
 	private String memberFood;
 
 	@Column
-	private double result;
+	private Integer result;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "member_id")
